requirements.txt
fastapi
uvicorn==0.18.2
python-dotenv==0.20.0
motor==3.0.0
pytz==2022.1
python-decouple
passlib
python-jose
email-validator
python-multipart
pytest-asyncio
requests
bcrypt